make filesystem depend on source + allow optional getSource field in fs
