+++ packages/components/src/lib/filesystem.ts
@@ -1,20 +1,5 @@
+import type { DirSource, FileKind, FileMetadata, FileSource, SourcePart } from './source.js'
 import { getFileName } from './utils.js'
-
-export type FileKind = 'file' | 'directory'
-
-export interface FileMetadata {
-  name: string
-  eTag?: string
-  size?: number
-  lastModified?: string
-  sourceId: string /// the source URL or path
-  kind: FileKind
-}
-
-export interface SourcePart {
-  text: string
-  sourceId: string
-}
 
 export interface FileSystem {
   fsId: string
@@ -25,6 +10,35 @@ export interface FileSystem {
   getResolveUrl: (sourceId: string) => string
   getSourceParts: (sourceId: string) => SourcePart[]
   listFiles: (prefix: string) => Promise<FileMetadata[]>
+  getSource?: (sourceId: string) => FileSource | DirSource
+}
+
+export function getSource(sourceId: string, fs: FileSystem): FileSource | DirSource | undefined {
+  if (fs.getSource) {
+    return fs.getSource(sourceId)
+  }
+  if (!fs.canParse(sourceId)) {
+    return undefined
+  }
+  const sourceParts = fs.getSourceParts(sourceId)
+  if (fs.getKind(sourceId) === 'file') {
+    return {
+      kind: 'file',
+      sourceId,
+      sourceParts,
+      fileName: fs.getFileName(sourceId),
+      resolveUrl: fs.getResolveUrl(sourceId),
+    }
+  } else {
+    const prefix = fs.getPrefix(sourceId)
+    return {
+      kind: 'directory',
+      sourceId,
+      sourceParts,
+      prefix,
+      listFiles: () => fs.listFiles(prefix),
+    }
+  }
 }
 
 function notImplemented(): never {
