+++ packages/components/src/lib/index.ts
@@ -1,11 +1,9 @@
-export { createHttpFileSystem, createHyperparamFileSystem } from './filesystem.js'
-export type { FileKind, FileMetadata, FileSystem, SourcePart } from './filesystem.js'
+export { createHttpFileSystem, createHyperparamFileSystem, getSource } from './filesystem.js'
+export type { FileSystem } from './filesystem.js'
 export { appendSearchParams, replaceSearchParams } from './routes.js'
 export type { RoutesConfig } from './routes.js'
-export { getSource } from './source.js'
-export type { DirSource, FileSource, Source } from './source.js'
+export type { DirSource, FileKind, FileMetadata, FileSource, Source, SourcePart } from './source.js'
 export { parquetDataFrame } from './tableProvider.js'
 export { asyncBufferFrom, cn, contentTypes, formatFileSize, getFileDate, getFileDateShort, getFileName, imageTypes, parseFileSize } from './utils.js'
 export { parquetQueryWorker } from './workers/parquetWorkerClient.js'
 export type { AsyncBufferFrom, Row } from './workers/types.js'
-
